core: move loaders to separate files
