+++ Blasphemous.LostDreams/LostDreams.cs
@@ -12,6 +12,7 @@ using Blasphemous.LostDreams.Items.QuestItems;
 using Blasphemous.LostDreams.Items.RosaryBeads;
 using Blasphemous.LostDreams.Items.SwordHearts;
 using Blasphemous.LostDreams.Levels;
+using Blasphemous.LostDreams.Levels.Loaders;
 using Blasphemous.LostDreams.Npc;
 using Blasphemous.ModdingAPI;
 using UnityEngine;
